Update consumer-client to use ZuAuth
