+++ apps/consumer-client/package.json
@@ -29,6 +29,7 @@
     "@pcd/webauthn-pcd": "0.10.0",
     "@pcd/zk-eddsa-event-ticket-pcd": "0.4.0",
     "@pcd/zk-eddsa-frog-pcd": "0.2.0",
+    "@pcd/zuauth": "0.0.1",
     "@personaelabs/spartan-ecdsa": "^2.1.4",
     "@semaphore-protocol/identity": "^3.15.2",
     "@simplewebauthn/browser": "^7.2.0",
