+++ apps/consumer-client/src/main.tsx
@@ -8,7 +8,7 @@ import GroupProof from "./pages/examples/group-proof";
 import SignatureProof from "./pages/examples/signature-proof";
 import ZkEDdSAEventTicketProof from "./pages/examples/zk-eddsa-event-ticket-proof";
 import ZkEDdSAFrogProof from "./pages/examples/zk-eddsa-frog-proof";
-import ZuAuthExample from "./pages/examples/zu-auth/example";
+import ZuAuth from "./pages/examples/zuauth";
 import Home from "./pages/Home";
 import ZupassPopupRedirect from "./pages/popup";
 
@@ -25,8 +25,8 @@ const router = createHashRouter([
     element: <ZkEDdSAEventTicketProof />
   },
   {
-    path: "examples/zu-auth",
-    element: <ZuAuthExample />
+    path: "examples/zuauth",
+    element: <ZuAuth />
   },
   { path: "examples/zk-eddsa-frog-proof", element: <ZkEDdSAFrogProof /> }
 ]);
