+++ apps/consumer-server/src/routing/routes/zuauth/generateWatermark.ts
