+++ apps/consumer-server/src/routing/routes/zuauth/logout.ts
